move the external-signer flags so they can be used by the voluntary exit command.
