+++ teku/src/main/java/tech/pegasys/teku/cli/options/ValidatorKeysOptions.java
@@ -13,10 +13,18 @@
 
 package tech.pegasys.teku.cli.options;
 
+import com.google.common.base.Strings;
+import java.net.MalformedURLException;
+import java.net.URL;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
+import org.apache.tuweni.bytes.Bytes;
 import picocli.CommandLine;
+import tech.pegasys.teku.bls.BLSPublicKey;
 import tech.pegasys.teku.config.TekuConfiguration;
+import tech.pegasys.teku.util.config.InvalidConfigurationException;
 
 public class ValidatorKeysOptions {
 
@@ -32,7 +40,81 @@ public class ValidatorKeysOptions {
       arity = "1..*")
   private List<String> validatorKeys = new ArrayList<>();
 
+  @CommandLine.Option(
+      names = {"--validators-key-files"},
+      paramLabel = "<FILENAMES>",
+      description = "The list of encrypted keystore files to load the validator keys from",
+      split = ",",
+      hidden = true,
+      arity = "0..*")
+  private List<String> validatorKeystoreFiles = new ArrayList<>();
+
+  @CommandLine.Option(
+      names = {"--validators-key-password-files"},
+      paramLabel = "<FILENAMES>",
+      description = "The list of password files to decrypt the validator keystore files",
+      split = ",",
+      hidden = true,
+      arity = "0..*")
+  private List<String> validatorKeystorePasswordFiles = new ArrayList<>();
+
+  @CommandLine.Option(
+      names = {"--validators-external-signer-public-keys"},
+      paramLabel = "<STRINGS>",
+      description = "The list of external signer public keys",
+      split = ",",
+      arity = "0..*")
+  private List<String> validatorExternalSignerPublicKeys = new ArrayList<>();
+
+  @CommandLine.Option(
+      names = {"--validators-external-signer-url"},
+      paramLabel = "<NETWORK>",
+      description = "URL for the external signing service",
+      arity = "1")
+  private String validatorExternalSignerUrl = null;
+
+  @CommandLine.Option(
+      names = {"--validators-external-signer-timeout"},
+      paramLabel = "<INTEGER>",
+      description = "Timeout (in milliseconds) for the external signing service",
+      arity = "1")
+  private int validatorExternalSignerTimeout = 1000;
+
   public void configure(TekuConfiguration.Builder builder) {
-    builder.validator(config -> config.validatorKeys(validatorKeys));
+    builder.validator(
+        config ->
+            config
+                .validatorKeys(validatorKeys)
+                .validatorExternalSignerPublicKeys(parseExternalSignerPublicKeys())
+                .validatorExternalSignerUrl(parseValidatorExternalSignerUrl())
+                .validatorExternalSignerTimeout(validatorExternalSignerTimeout)
+                .validatorKeystoreFiles(validatorKeystoreFiles)
+                .validatorKeystorePasswordFiles(validatorKeystorePasswordFiles));
+  }
+
+  private List<BLSPublicKey> parseExternalSignerPublicKeys() {
+    if (validatorExternalSignerPublicKeys == null) {
+      return Collections.emptyList();
+    }
+    try {
+      return validatorExternalSignerPublicKeys.stream()
+          .map(key -> BLSPublicKey.fromSSZBytes(Bytes.fromHexString(key)))
+          .collect(Collectors.toList());
+    } catch (IllegalArgumentException e) {
+      throw new InvalidConfigurationException(
+          "Invalid configuration. Signer public key is invalid", e);
+    }
+  }
+
+  private URL parseValidatorExternalSignerUrl() {
+    if (Strings.isNullOrEmpty(validatorExternalSignerUrl)) {
+      return null;
+    }
+    try {
+      return new URL(validatorExternalSignerUrl);
+    } catch (MalformedURLException e) {
+      throw new InvalidConfigurationException(
+          "Invalid configuration. Signer URL has invalid syntax", e);
+    }
   }
 }
